Optimize svg files
I manually made some changes to the chef logo to try to shrink down the
complexity of the svg, then tried playing around with some tools to
automate the process a bit further. I'm pulling in svgo and adding it
into the build, but accidentally applied it to the source files
themselves just now before committing.