Remove deprecated packages causing security alerts
The github security bot pointed out a number security vulnerabilities
that we weren't actually using, but that were installed when I started
to look into phantomjs as an option for generating some of our logos.
Apparently I never should have installed the phantomjs package itself,
but allow it's newer prebuilt image to get pulled in by some of the
other more active projects that are being maintained.